+++ lib/al/Library/Movement/EnemyStateBlowDown.h
@@ -1,14 +1,29 @@
 #pragma once
 
-#include <basis/seadTypes.h>
 #include <math/seadVector.h>
 
 #include "Library/Nerve/NerveStateBase.h"
 
 namespace al {
-struct EnemyStateBlowDownParam;
 class HitSensor;
 class LiveActor;
+
+struct EnemyStateBlowDownParam {
+public:
+    EnemyStateBlowDownParam();
+    EnemyStateBlowDownParam(const char* actionName);
+    EnemyStateBlowDownParam(const char* actionName, f32 velocityStrength, f32 gravityStrength,
+                            f32 velocityMultiplier, f32 velocityScale, s32 blowDownLength,
+                            bool faceAwayFromActor);
+
+    const char* mActionName = "BlowDown";
+    f32 mVelocityStrength = 10.3f;
+    f32 mGravityStrength = 28.2f;
+    f32 mVelocityMultiplier = 1.1f;
+    f32 mVelocityScale = 0.995f;
+    s32 mBlowDownLength = 120;
+    bool mFaceAwayFromActor = true;
+};
 
 class EnemyStateBlowDown : public ActorStateBase {
 public:
